% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONIapp.R
\name{filterResults}
\alias{filterResults}
\title{Filter graph data frames using user provided parameters}
\usage{
filterResults(
  Master,
  MasterRNA,
  ControlRNA,
  ControlCan,
  alt.data,
  Cancers = "all",
  Mutations = "all"
)
}
\arguments{
\item{Master}{A master dataframe from a RIGATONI output}

\item{MasterRNA}{RNA counts from a RIGATONI output}

\item{ControlRNA}{RNA counts from control samples within RIGATONI output}

\item{ControlCan}{Cancer type metadata for ControlRNA}

\item{alt.data}{Alteration dataframe from RIGATONI object}

\item{Cancers}{list of cancer types to analyze from the user}

\item{Mutations}{list of mutations to analyze from the user}
}
\value{
A list of dataframes for graph use
}
\description{
Filter graph data frames using user provided parameters
}
