% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONIapp.R
\name{getAltFunctionImmune}
\alias{getAltFunctionImmune}
\title{Recalcute alteration level function and immune status}
\usage{
getAltFunctionImmune(Master)
}
\arguments{
\item{Master}{A master dataframe from a RIGATONI output}
}
\value{
new alt.data data frame
}
\description{
Recalcute alteration level function and immune status
}
