% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONIapp.R
\name{getCancerTypeGraph}
\alias{getCancerTypeGraph}
\title{Get cancer type distribution}
\usage{
getCancerTypeGraph(Master, cancersFreq)
}
\arguments{
\item{Master}{master dataframe from RIGATONI output}

\item{cancersFreq}{cancer type frequency data}
}
\value{
ggplot2 object
}
\description{
Get cancer type distribution
}
