% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONIapp.R
\name{getExpressionBoxplot}
\alias{getExpressionBoxplot}
\title{Get gene level expression boxplots}
\usage{
getExpressionBoxplot(rna, metadata, gene, group)
}
\arguments{
\item{rna}{rna for graphs from RIGATONI output}

\item{metadata}{metadata for graphs from RIGATONI output}

\item{gene}{gene provided by user to analyze}

\item{group}{X axis parameter}
}
\value{
ggplot2 object
}
\description{
Get gene level expression boxplots
}
