% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONIapp.R
\name{getQuantFigures}
\alias{getQuantFigures}
\title{Get gene level expression boxplots}
\usage{
getQuantFigures(quant, metadata, group, includeOther = "Yes")
}
\arguments{
\item{quant}{quant data to use for analysis}

\item{metadata}{metadata for graphs from RIGATONI output}

\item{group}{X axis parameter}

\item{includeOther}{should the algorithm include "other" cell types in the result}
}
\value{
ggplot2 object
}
\description{
Get gene level expression boxplots
}
